requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
requests==2.31.0
prettytable==3.8.0